Build: (d45daa2) Improved downloadJaidScripts

+++ bin/downloadJaidScripts
@@ -34,8 +34,14 @@ for scriptRepo in "${scriptRepos[@]}"; do
     git -C "$repoFolder" checkout dist
   else
     printf "$stylePink%s already cloned, pulling$styleReset\n" "$scriptRepo"
-    git -C "$repoFolder" fetch origin dist
-    git -C "$repoFolder" reset --hard origin/dist
+    changeCount=$(git rev-list "HEAD...origin/$(git branch --show-current)" --count)
+    if [ "$changeCount" -gt 0 ]; then
+      printf "${stylePink}New commits: %s$styleReset\n" "$changeCount"
+      git -C "$repoFolder" fetch origin dist
+      git -C "$repoFolder" reset --hard origin/dist
+    else
+      printf "${stylePink}Up to date$styleReset\n"
+    fi
   fi
   repoBinFolder="$repoFolder/bin"
   find "$repoBinFolder" -type f -printf '%f\0' | while read -r -d $'\0' file; do
